Send the version number of the main system firmware module in the Hello message; minor fixes

+++ test/unit_tests/communication/protocol.cpp
@@ -31,7 +31,7 @@ class AbstractProtocol : public Protocol
 public:
 	AbstractProtocol(MessageChannel& channel) : Protocol(channel) {}
 
-	virtual size_t build_hello(Message& message, uint8_t flags)
+	virtual size_t build_hello(Message& message, uint16_t flags)
 	{
 		return 0;
 	}
